Refactored project structure.
