+++ .gitignore
@@ -3,8 +3,8 @@
 *.pyc
 *.exe
 *.toc
-*.pyc
 *.xml
 build/
-pdf_comb_az/
+dist/
 __pycache__/
+.idea
