+++ src/gui.py
@@ -21,7 +21,7 @@ def resource_path(relative_path):
         # PyInstaller creates a temp folder and stores path in _MEIPASS
         base_path = sys._MEIPASS
     except Exception:
-        base_path = os.path.abspath(".")
+        base_path = os.path.abspath("..")
 
     return os.path.normpath(os.path.join(base_path, relative_path))
 
@@ -39,18 +39,18 @@ class GUICore:
         self.OUTPUT_FILENAME = 'combined'
         self.TLF_METADATA_NAME = 'tlfmetadata'
         # Insert Atorus logo.
-        img1 = PhotoImage(file=resource_path(os.path.join('.img', 'atorus_logo.png')))
+        img1 = PhotoImage(file=resource_path(os.path.join('../assets/images', 'atorus_logo.png')))
         banner1 = Label(self.root, image=img1)
         banner1.image = img1
         banner1.place(x=40, y=4)
 
         # Insert PDF Utility logo.
-        img2 = PhotoImage(file=resource_path(os.path.join('.img', 'pdf_utility_logo.png'))).subsample(2, 2)
+        img2 = PhotoImage(file=resource_path(os.path.join('../assets/images', 'pdf_utility_logo.png'))).subsample(2, 2)
         banner2 = Label(self.root, image=img2)
         banner2.image = img2
         banner2.place(x=190, y=12)
         # Insert Python logo.
-        img3 = PhotoImage(file=resource_path(os.path.join('.img', 'python_logo.png'))).subsample(2, 2)
+        img3 = PhotoImage(file=resource_path(os.path.join('../assets/images', 'python_logo.png'))).subsample(2, 2)
         banner3 = Label(self.root, image=img3)
         banner3.image = img3
         banner3.place(x=405, y=12)
